+++ mlflow/projects/docker.py
@@ -10,10 +10,10 @@ import subprocess
 import docker
 
 from mlflow import tracking
-from mlflow.projects.utils import get_databricks_env_vars
 from mlflow.exceptions import ExecutionException
 from mlflow.projects.utils import MLFLOW_DOCKER_WORKDIR_PATH
 from mlflow.utils import process, file_utils
+from mlflow.utils.databricks_utils import get_databricks_env_vars
 from mlflow.utils.mlflow_tags import MLFLOW_DOCKER_IMAGE_URI, MLFLOW_DOCKER_IMAGE_ID
 from mlflow.utils.file_utils import _handle_readonly_on_windows
 from mlflow.utils.git_utils import get_git_commit
