+++ mlflow/projects/utils.py
@@ -5,10 +5,8 @@ import tempfile
 import urllib.parse
 import pathlib
 
-from distutils import dir_util
+from distutils import dir_util  # pylint: disable=W0402
 
-import mlflow.utils
-from mlflow.utils import databricks_utils
 from mlflow.utils.git_utils import get_git_repo_url, get_git_commit
 from mlflow.entities import SourceType, Param
 from mlflow.exceptions import ExecutionException
@@ -343,30 +341,3 @@ def get_run_env_vars(run_id, experiment_id):
         tracking._TRACKING_URI_ENV_VAR: tracking.get_tracking_uri(),
         tracking._EXPERIMENT_ID_ENV_VAR: str(experiment_id),
     }
-
-
-def get_databricks_env_vars(tracking_uri):
-    if not mlflow.utils.uri.is_databricks_uri(tracking_uri):
-        return {}
-
-    config = databricks_utils.get_databricks_host_creds(tracking_uri)
-    # We set these via environment variables so that only the current profile is exposed, rather
-    # than all profiles in ~/.databrickscfg; maybe better would be to mount the necessary
-    # part of ~/.databrickscfg into the container
-    env_vars = {}
-    env_vars[tracking._TRACKING_URI_ENV_VAR] = "databricks"
-    env_vars["DATABRICKS_HOST"] = config.host
-    if config.username:
-        env_vars["DATABRICKS_USERNAME"] = config.username
-    if config.password:
-        env_vars["DATABRICKS_PASSWORD"] = config.password
-    if config.token:
-        env_vars["DATABRICKS_TOKEN"] = config.token
-    if config.ignore_tls_verification:
-        env_vars["DATABRICKS_INSECURE"] = str(config.ignore_tls_verification)
-
-    workspace_info = databricks_utils.get_databricks_workspace_info_from_uri(tracking_uri)
-    if workspace_info is not None:
-        env_vars.update(workspace_info.to_environment())
-
-    return env_vars
