+++ mlflow/recipes/steps/evaluate.py
@@ -28,10 +28,9 @@ from mlflow.recipes.utils.tracking import (
     TrackingConfig,
     get_run_tags_env_vars,
 )
-from mlflow.projects.utils import get_databricks_env_vars
 from mlflow.protos.databricks_pb2 import INVALID_PARAMETER_VALUE
 from mlflow.tracking.fluent import _get_experiment_id, _set_experiment_primary_metric
-from mlflow.utils.databricks_utils import get_databricks_run_url
+from mlflow.utils.databricks_utils import get_databricks_env_vars, get_databricks_run_url
 from mlflow.utils.string_utils import strip_prefix
 
 _logger = logging.getLogger(__name__)
