+++ mlflow/store/artifact/databricks_artifact_repo.py
@@ -433,7 +433,7 @@ class DatabricksArtifactRepository(ArtifactRepository):
     def _parallelized_download_from_cloud(
         self, cloud_credential_info, file_size, dst_local_file_path, dst_run_relative_artifact_path
     ):
-        from mlflow.projects.utils import get_databricks_env_vars
+        from mlflow.utils.databricks_utils import get_databricks_env_vars
 
         try:
             parallel_download_subproc_env = os.environ.copy()
